# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/timezones": {
            "post": {
                "tags": [
                    "Timezones"
                ],
                "summary": "Insert timezones into the database",
                "description": "Inserts timezones into the database from an external API.",
                "operationId": "insertTimezones",
                "responses": {
                    "200": {
                        "description": "Timezones inserted successfully"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}